+++ jxls/src/main/java/org/jxls/builder/xls/XlsCommentAreaBuilder.java
@@ -12,7 +12,6 @@ import java.util.regex.Pattern;
 import org.jxls.area.Area;
 import org.jxls.area.CommandData;
 import org.jxls.area.XlsArea;
-import org.jxls.builder.AreaBuilder;
 import org.jxls.command.Command;
 import org.jxls.command.EachCommand;
 import org.jxls.command.GridCommand;
@@ -83,7 +82,7 @@ import org.jxls.util.LiteralsExtractor;
  *
  * @author Leonid Vysochyn
  */
-public class XlsCommentAreaBuilder implements AreaBuilder {
+public class XlsCommentAreaBuilder extends AbstractAreaBuilder {
     public static final String COMMAND_PREFIX = "jx:";
     private static final String ATTR_PREFIX = "(";
     private static final String ATTR_SUFFIX = ")";
@@ -137,72 +136,9 @@ public class XlsCommentAreaBuilder implements AreaBuilder {
 	public static Class<? extends Command> getCommandClass(String commandName) {
 		return commandMap.get(commandName);
 	}
-
-    /**
-     * Builds a list of {@link org.jxls.area.XlsArea} objects defined by top level AreaCommand markup ("jx:area")
-     * containing a tree of all nested commands
-     * 
-     * @param transformer -
-     * @param clearTemplateCells default is true
-     * @return Area list
-     */
-    @Override
-    public List<Area> build(Transformer transformer, boolean clearTemplateCells) {
-        List<Area> userAreas = new ArrayList<Area>();
-        List<CellData> commentedCells = transformer.getCommentedCells();
-        List<CommandData> allCommands = new ArrayList<CommandData>();
-        List<Area> allAreas = new ArrayList<Area>();
-        for (CellData cellData : commentedCells) {
-            String comment = cellData.getCellComment();
-            List<CommandData> commandDatas = buildCommands(transformer, cellData, comment);
-            for (CommandData commandData : commandDatas) {
-                if (commandData.getCommand() instanceof AreaCommand) {
-                    XlsArea userArea = new XlsArea(commandData.getAreaRef(), transformer);
-                    allAreas.add(userArea);
-                    userAreas.add(userArea);
-                } else {
-                    List<Area> areas = commandData.getCommand().getAreaList();
-                    allAreas.addAll(areas);
-                    allCommands.add(commandData);
-                }
-            }
-        }
-        for (int i = 0; i < allCommands.size(); i++) {
-            CommandData commandData = allCommands.get(i);
-            AreaRef commandAreaRef = commandData.getAreaRef();
-            List<Area> commandAreas = commandData.getCommand().getAreaList();
-            Area minArea = null;
-            List<Area> minAreas = new ArrayList<Area>();
-            for (Area area : allAreas) {
-                if (commandAreas.contains(area) || !area.getAreaRef().contains(commandAreaRef)) continue;
-                boolean belongsToNextCommand = false;
-                for (int j = i + 1; j < allCommands.size(); j++) {
-                    CommandData nextCommand = allCommands.get(j);
-                    if (nextCommand.getCommand().getAreaList().contains(area)) {
-                        belongsToNextCommand = true;
-                        break;
-                    }
-                }
-                if (belongsToNextCommand || (minArea != null && !minArea.getAreaRef().contains(area.getAreaRef()))) continue;
-                if (minArea != null && minArea.equals(area)) {
-                    minAreas.add(area);
-                } else {
-                    minArea = area;
-                    minAreas.clear();
-                    minAreas.add( minArea );
-                }
-            }
-            for (Area area : minAreas) {
-                area.addCommand(commandData.getAreaRef(), commandData.getCommand());
-            }
-        }
-        if (clearTemplateCells) {
-        	userAreas.forEach(area -> area.clearCells());
-        }
-        return userAreas;
-    }
     
-    private List<CommandData> buildCommands(Transformer transformer, CellData cellData, String text) {
+    @Override
+    protected List<CommandData> buildCommands(Transformer transformer, CellData cellData, String text) {
         List<CommandData> commandDatas = new ArrayList<CommandData>();
         List<String> commentLines;
         if (MULTI_LINE_SQL_FEATURE) {
